# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the Financial Partner application",
        "title": "Financial Partner API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the detailed information of the current user, with the option to return specific data fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "GetUser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Fields to include (profile, wallet, character). If not specified, returns all",
                        "name": "scope",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved user information",
                        "schema": {
                            "$ref": "#/definitions/dto.GetUserResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the current user's nickname",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "UpdateUser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update user request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update user successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/me/character": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the current user's character",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "UpdateUserCharacter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update user character request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUserCharacterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update user character successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.GetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CharacterResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "char_001"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/characters/advisor.png"
                },
                "name": {
                    "type": "string",
                    "example": "Character Name"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetUserResponse": {
            "type": "object",
            "properties": {
                "character": {
                    "$ref": "#/definitions/dto.CharacterResponse"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-03-07T12:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "60d6ec33f777b123e4567890"
                },
                "name": {
                    "type": "string",
                    "example": "User Name"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-03-07T12:00:00Z"
                },
                "wallet": {
                    "$ref": "#/definitions/dto.WalletResponse"
                }
            }
        },
...
        "dto.UpdateUserCharacterRequest": {
            "type": "object",
            "required": [
                "character_id",
                "image_url"
            ],
            "properties": {
                "character_id": {
                    "type": "string",
                    "example": "char_001"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/characters/advisor.png"
                }
            }
        },
        "dto.UpdateUserRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "New User Name"
                }
            }
        },
        "dto.UpdateUserResponse": {
            "type": "object",
            "properties": {
                "diamonds": {
                    "type": "integer",
                    "example": 100
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "60d6ec33f777b123e4567890"
                },
                "name": {
                    "type": "string",
                    "example": "New User Name"
                },
                "savings": {
                    "type": "integer",
                    "example": 5000
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-03-07T12:00:00Z"
                }
            }
        },
...
        "dto.WalletResponse": {
            "type": "object",
            "properties": {
                "diamonds": {
                    "type": "integer",
                    "example": 100
                },
                "savings": {
                    "type": "integer",
                    "example": 5000
                }
            }
        }
    }
}